Add async and defer tags to scripts
This will make the page load faster, as the browser will start rendering
the page while loading the JavaScript in the background.

Chose "defer" for the scripts because they might need a DOM tree and the
banner script returned by the banner server depends on the `stats.js`
file (containing global values for the banners).